+++ app/grandchallenge/anatomy/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from grandchallenge.anatomy.models import BodyRegion
+
+admin.site.register(BodyRegion)
