+++ app/grandchallenge/challenges/admin.py
@@ -2,7 +2,6 @@ from django.contrib import admin
 from django.contrib.admin import ModelAdmin
 
 from grandchallenge.challenges.models import (
-    BodyRegion,
     BodyStructure,
     Challenge,
     ChallengeSeries,
@@ -25,7 +24,6 @@ class ExternalChallengeAdmin(ModelAdmin):
 admin.site.register(Challenge, ChallengeAdmin)
 admin.site.register(ExternalChallenge, ExternalChallengeAdmin)
 admin.site.register(ChallengeSeries)
-admin.site.register(BodyRegion)
 admin.site.register(BodyStructure)
 admin.site.register(ImagingModality)
 admin.site.register(TaskType)
