+++ app/grandchallenge/challenges/filters.py
@@ -9,9 +9,9 @@ from django_filters import (
 )
 from django_select2.forms import Select2MultipleWidget
 
+from grandchallenge.anatomy.models import BodyRegion
 from grandchallenge.challenges.forms import ChallengeFilterForm
 from grandchallenge.challenges.models import (
-    BodyRegion,
     BodyStructure,
     Challenge,
     ChallengeSeries,
