+++ app/grandchallenge/core/management/commands/init_gc_demo.py
@@ -17,8 +17,8 @@ from userena.models import UserenaSignup
 
 import grandchallenge.cases.models
 from grandchallenge.algorithms.models import Algorithm, AlgorithmImage
+from grandchallenge.anatomy.models import BodyRegion
 from grandchallenge.challenges.models import (
-    BodyRegion,
     BodyStructure,
     Challenge,
     ChallengeSeries,
